\name{HMMEMdirTrans-package}
\alias{HMMEMdirTrans-package}
\alias{HMMEMdirTrans}
\docType{package}
\title{
\packageTitle{HMMEMdirTrans}
}
\description{
\packageDescription{HMMEMdirTrans}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HMMEMdirTrans}
\packageIndices{HMMEMdirTrans}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{HMMEMdirTrans}

Maintainer: \packageMaintainer{HMMEMdirTrans}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
